feat(api): initial api implementation + glibc
Upon testing the API we noticed that when the app is compiled with musl
the API would perform roughly 80% worse than the glibc counter part.
We will have to do real benchmarks to see where this performance drop
off is coming from and likely have to ticket it to the musl / rust /
tokio devs.