feat: add terraform deploy
Added auto deploy using terraform.
Terraform is a tool used to handle infrastructure state.
We have our server infra state stored in
https://github.com/ScuffleTV/terraform and this commit is for our
deployment specs for our applications. Terraform is very powerful and we
use terraform cloud and remote execution to run all our jobs on
Terraform cloud (which is entirely free). Terraform also supports
workspaces so we can easily add a stage and development environment in
the future, however we didn't for this PR.

fixes #28