feat: gql pubsub event subscriptions with redis
This PR adds a redis subscriber which manages redis pubsub
subscriptions. Aswell as lays the foundation for GQL subscriptions by
adding a username displayname update subscription event. However there
is currently no producer of this event, that will come when we add an
API to be able to change your display name and username.

Resolves #87 point 2