feat: improve player (#85)
- Added ABR (#85 point 1)
Adaptive bitrate streaming is the ability for the player to dynamically
change quality based on network conditions
- Added Codec Support (#85 point 3)
We now check if the browser supports AV1 / HEVC / OPUS before we list the
quality as a potential variant.
- Added event manager to handle events from the player.

Note: We added some unsafe code. You should read the safety comment on
what it is and how it works.