chore(ci): update dependencies and fix website lint
Updates all deps for JS and Rust and fixes lint and build for website
docker.

On website, switching to the latest version of urlq caused us to redo
our design for the custom exchange that forwarded all requests to the
websocket (if open) otherwise all requests to rest. With the new version
of urlq the way that it is implemented now is the way you are supposed
to do this.

Improve CI Build process and test running