feat: complete video services
This commit aims to complete edge, transcoder, ingest and other video
libs.

The remaining work is left on video api.

This is being merged prior to video api being complete due to the
addition of the dataloader in the common crate which will provide use to
the platform api work.