feat: migrate to common::dataloader
This commit migrates to the new dataloaders and also takes the work to
move the old db struct style to the new style. Syncing with the video
side.

Remaining work would be to update the protobuf files for events to stop
using strings and make use of the Ulid proto.