fix: remove openssl
This commit switches to using rustls over openssl.
Unfortunately, rust-jwt needs openssl for the PKeyVerify impl, so we
also forked their repo to add support for openssl-vendored which means
that it will be staticly linked rather than a system-dep. This feature
is only used by video-edge and lto-fat purges it from all other
binaries.