feat: use generic global traits
The reason behind this is because we ideally should be able to provide a
single binary to launch the entire system. This also reduces the
code duplication across the binaries.

This PR also introduces changes made to
https://github.com/async-graphql/async-graphql/pull/1424 on the
platform-api. We introduced generics into async-graphql unions so that
we can make use of our generic global approach.

This PR also introduced a #[config(flatten)] directive which allows for
config substructs to be flattened into the parent.