feat: remove temporary folder from image processor
With the new addition of the `ffmpeg-next-io` crate we can now do
in-memory io with `ffmpeg-next`. Previously we needed to write the file
to a temporary location and then clean it up after the job was complete.
Now we can just buffer the entire file in memory and process it in
place.