chore: fixed paths for syncing labels (#181)
* chore: fixed paths for syncing labels

Now the workflow for the `Sync Lables` is corectly configured.

For faster performance, we configured the **fetch-depth** option
and set the depth to 1. This allows for faster pulling of the
repository, resulting in improved performance.

* style: no need double quotes color hex in uppercase

The color values on GitHub pages for adding labels are always in
uppercase. The original color has not changed, it remains the same.

Double quotes for the name and description are not needed in this
file, they have been removed.

* style: sort label name ascending

The labels are sorted in ascending order within the particular
labels group, which makes it easier to locate a specific label.