feat: cancel task threads
Add a new feature which allows for cancel blocking tasks on threads.
This allows for long blocking tasks (encoding) to yield back to some
checker which unwinds the stack if the task has been aborted.
This only works if you call the checker, but for most cases we dont have
long running code in C or some other library. So its perfectly fine and
fixes a lot of the problems we have with encoding blocking shutdown on
`tokio::spawn_blocking`.