fix(lint): clippy bug
Fix a bug introduced in the latest release of nightly clippy. Clippy
believes that the `.as_ref()` is redundant here but it is not. What it
proposes is to use a `.clone()` since `.as_ref().map(..)` followed by a
clone in the `.map(..)` is "the same as" `.clone()` however the bug is the fact
we are mapping to a different value and then cloning, without mapping we
would be cloning the full struct, rather than only the part we want.