chore: change license
The reason for this change is that we want to license all library code
under MIT or Apache 2.0 and then license ScuffleVideo as BSL-1.0.
The BSL-1.0 License will convert to Apache 2.0 after 4 years. The BSL
license allows for anyone to use the software baring the fact that they
are not running a managed service of Scuffle Video to sell that to 3rd
parties.