fix: remove jemalloc
We have observed strange behavior with jemalloc causing segfaults within
the image processor library. We ran the library with libasan and
valgrind (both incompatiable with jemalloc) and we could not find any
leaks or invalid memory access or double frees. However when we use
jemalloc we have issues and sometimes it segfaults always deep inside
jemalloc code. Since we cannot debug this with valgrind / libasan its
impossible to figure out if its a problem in our code (more likely) or
jemalloc (unlikely but possible), given that the bug only comes up when
we use jemalloc its likely a difference in strictness of the allocator
and something we are doing is wrong but the system allocator doesnt care
but jemalloc does.

This issue is also only present when we link directly to jemalloc + use
jemalloc as the global rust allocator (so it might be an issue in the
tikv-jemalloc crate). When we use jemalloc via `LD_PRELOAD` the issue
isnt there.